+++ gen/quaycrew/v1/controlplane_grpc.pb.go
@@ -31,6 +31,7 @@ const (
 	ControlPlaneService_StartFlow_FullMethodName               = "/quaycrew.v1.ControlPlaneService/StartFlow"
 	ControlPlaneService_GetFlowRun_FullMethodName              = "/quaycrew.v1.ControlPlaneService/GetFlowRun"
 	ControlPlaneService_ListFlowRuns_FullMethodName            = "/quaycrew.v1.ControlPlaneService/ListFlowRuns"
+	ControlPlaneService_StopFlowRun_FullMethodName             = "/quaycrew.v1.ControlPlaneService/StopFlowRun"
 	ControlPlaneService_AttachChannel_FullMethodName           = "/quaycrew.v1.ControlPlaneService/AttachChannel"
 	ControlPlaneService_SetSecret_FullMethodName               = "/quaycrew.v1.ControlPlaneService/SetSecret"
 	ControlPlaneService_ListSecrets_FullMethodName             = "/quaycrew.v1.ControlPlaneService/ListSecrets"
@@ -73,6 +74,7 @@ type ControlPlaneServiceClient interface {
 	StartFlow(ctx context.Context, in *StartFlowRequest, opts ...grpc.CallOption) (*StartFlowResponse, error)
 	GetFlowRun(ctx context.Context, in *GetFlowRunRequest, opts ...grpc.CallOption) (*GetFlowRunResponse, error)
 	ListFlowRuns(ctx context.Context, in *ListFlowRunsRequest, opts ...grpc.CallOption) (*ListFlowRunsResponse, error)
+	StopFlowRun(ctx context.Context, in *StopFlowRunRequest, opts ...grpc.CallOption) (*StopFlowRunResponse, error)
 	AttachChannel(ctx context.Context, in *AttachChannelRequest, opts ...grpc.CallOption) (*AttachChannelResponse, error)
 	SetSecret(ctx context.Context, in *SetSecretRequest, opts ...grpc.CallOption) (*SetSecretResponse, error)
 	ListSecrets(ctx context.Context, in *ListSecretsRequest, opts ...grpc.CallOption) (*ListSecretsResponse, error)
@@ -221,6 +223,16 @@ func (c *controlPlaneServiceClient) ListFlowRuns(ctx context.Context, in *ListFl
 	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListFlowRunsResponse)
 	err := c.cc.Invoke(ctx, ControlPlaneService_ListFlowRuns_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *controlPlaneServiceClient) StopFlowRun(ctx context.Context, in *StopFlowRunRequest, opts ...grpc.CallOption) (*StopFlowRunResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(StopFlowRunResponse)
+	err := c.cc.Invoke(ctx, ControlPlaneService_StopFlowRun_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -465,6 +477,7 @@ type ControlPlaneServiceServer interface {
 	StartFlow(context.Context, *StartFlowRequest) (*StartFlowResponse, error)
 	GetFlowRun(context.Context, *GetFlowRunRequest) (*GetFlowRunResponse, error)
 	ListFlowRuns(context.Context, *ListFlowRunsRequest) (*ListFlowRunsResponse, error)
+	StopFlowRun(context.Context, *StopFlowRunRequest) (*StopFlowRunResponse, error)
 	AttachChannel(context.Context, *AttachChannelRequest) (*AttachChannelResponse, error)
 	SetSecret(context.Context, *SetSecretRequest) (*SetSecretResponse, error)
 	ListSecrets(context.Context, *ListSecretsRequest) (*ListSecretsResponse, error)
@@ -534,6 +547,9 @@ func (UnimplementedControlPlaneServiceServer) GetFlowRun(context.Context, *GetFl
 }
 func (UnimplementedControlPlaneServiceServer) ListFlowRuns(context.Context, *ListFlowRunsRequest) (*ListFlowRunsResponse, error) {
 	return nil, status.Error(codes.Unimplemented, "method ListFlowRuns not implemented")
+}
+func (UnimplementedControlPlaneServiceServer) StopFlowRun(context.Context, *StopFlowRunRequest) (*StopFlowRunResponse, error) {
+	return nil, status.Error(codes.Unimplemented, "method StopFlowRun not implemented")
 }
 func (UnimplementedControlPlaneServiceServer) AttachChannel(context.Context, *AttachChannelRequest) (*AttachChannelResponse, error) {
 	return nil, status.Error(codes.Unimplemented, "method AttachChannel not implemented")
@@ -834,6 +850,24 @@ func _ControlPlaneService_ListFlowRuns_Handler(srv interface{}, ctx context.Cont
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ControlPlaneServiceServer).ListFlowRuns(ctx, req.(*ListFlowRunsRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _ControlPlaneService_StopFlowRun_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(StopFlowRunRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ControlPlaneServiceServer).StopFlowRun(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: ControlPlaneService_StopFlowRun_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ControlPlaneServiceServer).StopFlowRun(ctx, req.(*StopFlowRunRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -1288,6 +1322,10 @@ var ControlPlaneService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "ListFlowRuns",
 			Handler:    _ControlPlaneService_ListFlowRuns_Handler,
+		},
+		{
+			MethodName: "StopFlowRun",
+			Handler:    _ControlPlaneService_StopFlowRun_Handler,
 		},
 		{
 			MethodName: "AttachChannel",